requirements.txt
streamlit
google-genai
pillow
pandas
supabase